updated setup.py, and related sundries, to be python 3 friendly
Changed setup.py so that it can run both under 2.7 and >=3.5. Now it can attempt
build and miserably fail on C/C++ compilations.
Introduced a new dependency, on 'future', for python2.7.
We now state that we support both 2.7 and 3.5.

Few compatibility changes, all localized on python files that are loaded by
setup.py.

+++ pydoop/hdfs/path.py
@@ -24,9 +24,8 @@ pydoop.hdfs.path -- Path Name Manipulations
 import os
 import re
 import time
-
+from past.builtins import long
 from . import common, fs as hdfs_fs
-
 
 curdir, pardir, sep = '.', '..', '/'  # pylint: disable=C0103
 
@@ -48,7 +47,7 @@ class StatResult(object):
     def __init__(self, path_info):
         self.st_mode = path_info['permissions']
         self.st_ino = 0
-        self.st_dev = 0L
+        self.st_dev = long(0)
         self.st_nlink = 1
         self.st_uid = path_info['owner']
         self.st_gid = path_info['group']
